Make substitution engine ignore factor arrays, fix #302
The fix assumes, that the bug was that the substitution engine -
tox.config.Replacer, while recursvely replacing references to actual
values, treated factor arrays (like {py27,py34}) as references, that
are needed to be replaced, which they are not.

The patch simply amends regex which is used by Replacer to recognize
references, to ignore factor arrays. And factor arrays assumed to be
values, enclosed in curly braces, that contain at least one comma.

--HG--
branch : issue302