use xfail_strict and adapt affected tests (#588)
We're using tests marked as xfail to reproduce known issues (e.g. upstream bugs), but at the moment we do not fail a test, when it passed unexpectedly. Adding xfail_strict = True to the pytest configuration makes sure we notice if a bug was "accidentally" fixed or fixed in upstream.