run install_commands from a `{toxinidir}/install_from_here` folder
This avoids any adverse interactions between the current folder and the
installation command, e.g. `python -m pip` command will add the current
folder to the Python path, and thus fail to install any package located
in that folder for which there also exists an `.egg-info` folder in that
same folder.