Add a -q (quiet) option to tox (#682)
* Add a -q (quiet) option to tox
Sometimes the output of tox in addition to the output of the command(s)
tox is running can be a overwhelming and confusing. Allowing users to
silence tox's output is incredibly useful in focusing on the
command(s)'s output. Resolves #256
* Add a changelog item and add myself to contributors
* Rephrase our changelog item and fixup change
Use a class to contain named constants to replace all the magic numbers
used in the reporter. Further, rename the parsed values from the
command-line.