spring cleaning / refactoring
Sorry, this is a lot and it will get even more. I have some large chunks of time to spent with tox and I am going through with the big brush. I want to move towards a big clean up which will result in a tox 4 not to far into the future. These are the first steps. Part of this is getting stuff sorted for plugins to be more like 1. class citizens and stabilizing the API. A lot of things have been marked as experimental for many years now and it is time to let this go and rather deprecate things, if they should go away, instead of letting them hang around as experimental for ages.

Start hookimpl/hookspec objects deprecation

They are used in some places by importing direclty from tox. This might have been the right way a loing time ago, but now pluggy has an API to create those objects via pluggy.HookImplMarker and pluggy.HookSpecMarker without the need of importing them.

hookimpl needs to be initiated for internal use and should be marked that way. Plugin writers
should use the official API instead of importing it from tox.

* removed all experimental markers from docstrings. If they turn out wrong/bad,
  we just have to deprecate them and remove them in a major release
* also hookimpl is instantiated twice - mark these objects as deprecated and
  get rid of one of them later
* especially having a hookimpl object in the hookspecs module is unnecessary and confusing
* access hookimpl as module attribute of tox (clearer where it comes from)
* removed some redundant rst references. If they are initialized once in any
  namespace they are defined globally (I don't like it either but that is how it
  works in restructuredText and ignroring that and defining the same things
  several times is confusing)

tox-quickstart script and tests overhaul

* never overwrite existing files - even if the user tries
* make py36 the default environment rather than py27 if 1 is chosen
* do not refer to the configuration as tox.ini as it might not be tox.ini but something else
* refactor tests for readability and maintainability
* adapt tests to new behaviour
* refine some code having to do with usere input that turns into lists
* remove unused flags from main
* naming things (like, too terse or shadowing builtins, etc.)
* whitespace (try to make it more consistent)
* remove some redundant internal tests. Going through main is perfectly adequate and makes the tests more resilient to internal changes

Other stuff

* tox has an API already although this was never really planned - add comment to __all__ to make clear that this is the only 'official' API and that hookimpl and hookspec are not part of it
* tidy up imports (prefer module imports to make clear where stuff comes from)
* move _dummy object to only class where it is needed and name it like the constant that it is
* use six for py2/3 compatibility where I notice it and where it is cheap to do
* remove dead code
* move exception related code into its own module rather than having it hang around in __init__.py
* there is no need to have an intermediate run_main - that is used everywhere only under the name tox.cmdline rename function and use it directly, just like client code is using it
* add a fixture for working in a clean tmpdir