handle interpreter directives with long lengths (#794)
* venv: prepend interpreter directive to argument list

When preparing virtual environments in a file container which has
large length, the system might not be able to invoke shebang scripts
which define interpreters beyond system limits (e.x. Linux as a
limit of 128; BINPRM_BUF_SIZE [1]). This method can be used to check
if the executable is a script containing a shebang line [2]. If so,
extract the interpreter (and possible optional argument) and prepend
the values to the provided argument list. tox will only attempt to
read an interpreter directive of a maximum size of 2048 bytes to
limit excessive reading and support UNIX systems which may support a
longer interpret length.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/tree/include/uapi/linux/binfmts.h?h=linux-4.16.y#n19
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-stable.git/tree/fs/binfmt_script.c?h=linux-4.16.y#n24

Signed-off-by: James Knight <james.d.knight@live.com>

* tests: validate shebang interpreter prepending

Providing a series of test assertions to validate the recently added
method `prepend_shebang_interpreter`. Ensures shell scripts with various
interpreter directives (whitespaces, single argument, invalid entries)
are appended or ignored respectfully.

Signed-off-by: James Knight <james.d.knight@live.com>

* changelog: track pr 794

Track change which supports handling interpreter directives in
environments with long path lengths [1].

[1]: e913dd3f81692c8e8432354a097a51551a9aa160

Signed-off-by: James Knight <james.d.knight@live.com>

* doc: add workaround for limited shebang environments

Adding documentation to tox's configuration document describing how
to take advantage of the `TOX_LIMITED_SHEBANG` environment variable
to bypass system-defined shebang interpreter directive limits.

Signed-off-by: James Knight <james.d.knight@live.com>