Optimize CI (#876)
Don't run tests in parallel on Travis. Running the test suite in parallel
on Travis does not speed it up, so drop it. It aslo crashes pypy runs.

Move sudo required and dist xenial to base.

Remove testing against Python nigthly. Testing against dev is enough.

Skip the deploy stage entirely if not a PR.

Bump pre-commit version.

Appveyor build is only required if we modified the source files or the
build config.