Touch '.egg-info' directory after install (#909) (#910)
When a project is using `skipsdist = True` and `usedevelop = True`, tox will attempt to determine when it needs to reinstall the package (the `develop-inst-nodeps` activity) and when it does not (the `develop-inst-noop` target). It does this by comparing the last modified timestamps of the 'setup.py' or 'setup.cfg' files with that of the `.egg-info` directory.

Unfortunately, this is a flawed check as, even after running the `develop-inst-nodeps` activity (which executes a variant of `pip install -e $setupdir --no-deps`) the timestamp of the `.egg-info` directory is not updated. This is a particularly irritating issue for larger projects when the install step can take many seconds or even minutes.

The ideal fix would be in setuptools. However, while we wait on this, we can fix it ourselves by simply `touch`ing the `.egg-info` directory after a reinstall. We do this.