Avoid build package becoming unavailable during session run
We create a hard link (if possible, otherwise copy of the
built package) inside a temp folder. This way subsequent sessions are
free to remove dist folder and we also guarantee that the built package
during the build is what tox environments use (even if the package
gets rebuilt in the meantime).