Allow user to inject specific Pythons to try as step 1 during discovery
At the moment it's unecessarily hard to specify a specific python
executable to use for a given spec. While mangling with the path
might work for environments for tox environments that require a python
differing the host tox python version; for environments matching the
host tox interpreter it's impossible to use any other interpreter than
what tox uses. Here we add a CLI argument (fallback to some environment
variable for CI usage) that allows the user to inject custom python
interpreters as step 1 of the discovery mechanism.

Signed-off-by: Bernat Gabor <bgabor8@bloomberg.net>