Do not use setup.cfg file that does not have tox:tox namespace
When tox is invoked without a tox configuration file (and without command line
options that could be used instead), it fails with:

    $ tox
    ERROR: tox config file (either pyproject.toml, tox.ini, setup.cfg) not found

This is useful: If there is no config, test didn't run and such error should
not pass silently.

However, until this commit, when there was a setup.cfg file it was considered
as a tox config even when there was no [tox:tox] section:

    $ tox
    GLOB sdist-make: .../setup.py
    python create: .../.tox/python
    python inst: ...
    python installed: ...
    python run-test-pre: PYTHONHASHSEED='1234567890'
    ___________________________________ summary ____________________________________
      python: commands succeeded
      congratulations :)

That is not safe, because setup.cfg can exist even if there is no tox
configuration in it and the success creates a false sense of safety.

To fix this and to match the behavior of pyproject.toml without a tox section,
tox now skips setup.cfg, if there is no [tox:tox] section.

Fixes https://github.com/tox-dev/tox/issues/1045