Use pytest-cov instead of multiple coverage.py calls
This speeds up a bit the run, because we don't have to pay the python
and coverage process startup time five times, but just once. Also,
change the tox test environments arguments to fallback to just pytest
when calling it via posargs (in these cases we're not interested in
coverage).

Signed-off-by: Bernát Gábor <bgabor8@bloomberg.net>