Requirement files support within deps
To support -r/-c/etc within deps we consider the deps section itself a
requirements file. We forward this as such to the installer. We validate
and traverse the entries within the requirements file, build the keys
so we know when we need to rebuild the tox environment. If only add
operation has been observed do not recreate the environment, bur rather
just update it with a new requirement file install.

Also added a mechanism into tox.pytest that allows enabling a short
circuit into the execute API. This is useful for example to auto
pass all install steps with success and a noop.

All install steps now have a run id starting with install. Run ids
have been altered to use the _ over the - as separator. This is to
bring it inline with how the PEP-517 backend uses the function name as
run it, and as such uses the _ as separator.

Fold legacy dev build requires package install into install-deps from a
dedicated step.

Signed-off-by: Bernát Gábor <bgabor8@bloomberg.net>