venv: do not prepend a truncated shebang interpreter
When a user invokes a `tox` call with `TOX_LIMITED_SHEBANG` set, any
shebang-based scripts will have their interpreter's extracted and
explicitly executed. An issue with the initial implementation is that if
the configured interpreter is truncated to a path of another executable,
the wrong executable can be invoked instead (see also [1][2]). To
prevent this, an extra byte is read to ensure the length of the
extracted interpreter is within the configured limit (MAXINTERP);
otherwise the call will know the interpreter has been truncated and will
fallback onto the original argument list.

[1]: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/fs/binfmt_script.c?id=8099b047ecc431518b9bb6bdbba3549bbecdc343
[2]: https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/fs/binfmt_script.c?id=b5372fe5dc84235dbe04998efdede3c4daa866a9

Signed-off-by: James Knight <james.d.knight@live.com>