Fix CTRL+C not working on Windows
Seems sending CTRL_C event on Windows to a new process group does not
work in general. It registers on cmd.exe and powershell.exe on its own,
however is ignored on powershell when ran within Windows Terminal or
Wezterm. Sending CTRL_BREAK seems to work, but on the later brings down
the shell itself.

However, Windows works differently with signals than Linux. On Linux
only the root process gets the signal. On Windows when a signal is sent
to a process all processes within that process group will get an
instance of that signal. This means that on Windows we don't need to
forward signals. All child automatically gets it. The only thing we do is
escalate the unanswered interrupt to process kill (there's no SIGTERM on
Windows).

Signed-off-by: Bernát Gábor <gaborjbernat@gmail.com>