Do not reinstall in develop mode without setup.py
The detection logic for figuring out whether a editable install needs to
be reinstalled only works for setuptools based builds using setup.py.
If setup.py cannot be found the code will fail with an invocation error.
Currently no standard mechanism exists to query a package installed in
editable mode whether it should be reinstalled or not. Hence, this
change skips reinstallation if setup.py cannot be found.

Fixes #2197