venv: Do not fail for test commands that don't exist with "-" leader
Commands which fail, even at (for example) dynamic loader stage get a
pass if there's a "-" leader to indicate to ignore errors. That means
that any user using this must accept that there's a chance that the
command will not execute whatsoever.

Since these outcomes are the same as the executable not existing at all,
allow "-" to also silence the command not existing in the first place.

As an example, with:

    [testenv:bogus]
    commands =
	-bogus

Before:

    ERROR: InvocationError for command could not find executable bogus
    ________________________ summary ________________________
    ERROR:   bogus: commands failed

After:

    WARNING: command not found but explicitly ignored
    cmd: bogus
    ________________________ summary ________________________
    bogus: commands succeeded
    congratulations :)

Closes #2315.