Delete `tox` mailing list
After careful consideration, we, the tox maintainers, have come to the
conclusion that it is not worth any more to continue the tox mailing list.

In the last year, and especially in the last couple of months there have been
almost only spam emails, which are both distracting and bind maintainer time
(block spammers etc).

As alternative channels to contact us and to get help, we have

- https://github.com/tox-dev/tox/discussions
- https://discord.gg/tox
- https://stackoverflow.com/questions/tagged/tox

and all maintainers are active on Twitter, too.