Fix encoding used to read our own log files
When running Tox on a French Canadian Windows computer under an
account with a username that contains diacritics, Tox crashes with a
`UnicodeDecodeError` because it tries to read its log files with a
hard-coded encoding of `UTF-8`.  These log files contain output from
`python -m virtualenv` and `python -m pip`, whose output contains
references to the username (such as the contents of the ``APPDATA``
environment variable).  The problem is that the subprocesses which
open log files without any explicit encoding.  In that case, the
built-in `open()` uses `locale.getpreferredencoding(False)`, which is
`"cp1252"`.  A username containing an "Latin Small Letter E with
Acute" will be encoded as `\xe9`, which is not valid UTF-8 (the valid
UTF-8 sequence is `\xc3\xa9`).

One workaround is to run Tox with Python in "UTF-8 Mode".  This can be
achieved by setting the `PYTHONUTF8=1` environment variable or by
calling with `python -Xutf8 -m tox`.  Unfortunately, this is not in
the Tox documentation and can be quite confusing.  I spent more than
two hours troubleshooting this (I develop with Python full time and
I'd never heard of the "UTF-8 mode" before).

Of course, we could document the workaround, but it just seems good to
re-open the log files using the same encoding they were opened with in
the first place.

With this change, Tox works nicely with or without the "UTF-8 Mode".