Force virtualenv and pip to output UTF-8
After running these two commands with output redirected to a file, we
read said files with an explicit request for UTF-8 encoding.  On some
cases this does not contain UTF-8 data.  For example, on a French
Canadian Windows machine with non-ASCII characters in the username,
the output contains the contents of the `APPDATA` environment
variable, which is printed to the current code page (`cp1252`).

This change enables Python's "UTF-8 Mode" when running `virtualenv`
and `pip` to guarantee the output is UTF-8 everywhere.