Double backslash should escape to backslash
The resulting escaped value should not then be
subsequently used as an escape in future replacement
rounds.

Additional friends, like triple backslash, which
should escape the following, and quad backslash,
which should escape to two single backslashes
and not escape the following characters.

The Quad backslash test case also has escaped
double backslash at the beginning and end of the
string for more edge coverage.