Introduce Upgrading doc page for tox4 (#2805)
This adds a new page to the docs, Upgrading, meant to remove
the documentation regarding backwards incompatible changes in tox4
from the FAQ into a dedicated page.

The content of the page is primarily a lift-and-shift of the breaking
changes section of the FAQ, and the FAQ page and Upgrading pages now
cross-link to one another.

New content is added to the Upgrading page in the form of a preamble
for the doc.